# api/api.yaml
openapi: 3.0.0
info:
  title: Distributed Arithmetic Expression Calculator
  version: 1.0.0
servers:
  - url: http://localhost:8099
paths:
...
  /expressions/{id}:
    get:
      summary: Get expression
      operationId: getExpression
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Got expression
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expression'
        404:
          description: Expression not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        400:
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Expression:
      type: object
      properties:
        id:
          type: integer
        expression:
          type: string
        answer:
          type: string
        status:
          type: string
          enum: [completed, in process, invalid]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - expression
        - answer
        - status
        - createdAt
...
    Error:
      type: object
      properties:
        description:
          type: string
        id:
          type: number
      required:
        - description